build: update webpack and run npm audit fix [EE-761]
This fixes the critical security issue and a few modarate/high ones.

The next would be to update `gulp` to 5.0.0 to remove about half of the reported vulnerabitilites. I tried it and it seemed to work mostly fine except that the extension icon disappeared.  
We should tackle his in its own MR imho.

See eyeo/extensions/extensions!14