fix: Fix premium push notification prevention [EE-790]
The changes in this MR fix a JavaScript exception that occurs with the Premium Push notification preventer. The error was caused by the extension's migration to MV3.

This MR fixes the JavaScript exception and contains the code to update the two content scripts to TypeScript (and creates two new webpack bundles).

## Testing hints

To enable the Premium Push notification preventer:

* become a Premium user
* subscribe to the Distraction control filter list

See eyeo/extensions/extensions!11