feat: UI components initial structure [EXT-75]
This MR adds the `ui-components` folder to the monorepo. This folder is intended to host common UI components used by fragments and hosts.

I added the support for `React` in Adblock and the necessary build configuration for the components to be used directly.

To test it out, run the following commands in the Adblock host:
```sh
npm i
npm run build:dev:chrome
```
You should see a button in the popup UI, when clicked it should increase the counter.

:bangbang: Any suggestion on how to add types to this? Should we use typescript from the beginning?

See eyeo/extensions/extensions!20