fix: Move the 'received' event recording [IM-108]
The change in this MR moves the 'received' event recording for the `create_on_page_dialog` IPM command only to occur when the command is initially received.

## Testing hints

The current implementation of the `create_on_page_dialog` IPM command records a 'received' event when:
1. the command is received from the IPM server
2. the browser is restarted 
3. the service worker is restarted (MV3 extensions only)

With this update, the `create_on_page_dialog` IPM command will record a 'received' event when the command is received from the IPM server.  Verify that a browser or service worker restarted (MV3 extensions only) doesn't trigger the 'received' event recording.

In AdBlock, the current IPM events can be viewed in the debug data or with the following command on the service worker/background page:
`Prefs.ipm_events`

See eyeo/extensions/extensions!24