ci: Refined monorepo pipeline rules [EXT-73]
This fixes some problems found with the pipelines. 

We now specifically consider "changes" as changes in comparison with the `main` branch.

That means pipelines will run for a host if there are any changes to that host in your feature branch, not just for the last commit. 

We always run the full pipeline for the main branch and tags.

See eyeo/extensions/extensions!34