test: Skip uninstall tests on Edge MV2 [Refs EXT-153]
## Background

My initial intention was to file a MR which would retry the uninstall tests. However that looks complex: those tests can’t be simply retried because the extension needs to be re-installed to a new browser instance before that can happen.

One way to achieve that would be to re-run (not retry) failing test suites as described here: https://webdriver.io/docs/wdio-rerun-service/
However I’m not sure it’s worth the effort, because that implies using a new external package plus creating/maintaining shell scripts for a solution that would only be a workaround. So I’d rather skip the uninstall tests on Edge MV2, because that scenario will still be tested on Firefox MV2 and Chrome MV3.

## Changes

- Skips the uninstall tests on Edge MV2
- Removes the `connect ECONNREFUSED` try/catch mechanism because it's no longer needed after skipping the tests on Edge

See eyeo/extensions/extensions!40