docs: Standardising the release notes process [Refs EXT-160]
This adds documentation for how we intend to handle the release notes process.

## Testing note

This is a mostly internal process change. However, it does change the format of the release notes, which appear in the Adblock extension under the "Support" tab on the options page. Look for the heading "See the latest updates in our changelog".

See eyeo/extensions/extensions!33