feat: Add adjusted stylesheets [EXT-76]
This adds the stylesheets that have been deduplicated and sorted as part of !35 to the host. This will be useful as we implement new components and will keep the styles in sync, even as we slowly delete these items and apply the themes via Tailwind. ~~It does however need a quick visual review to be sure nothing has been rendered totally illegible in the consolidation, which is why it is currently marked as draft.~~ Issues identified in a visual review per theme have been rectified.

See eyeo/extensions/extensions!36