feat: Add local Dialogs that are triggered by auto allowlisting YT [EE-690]
The changes in this MR include the following:
- Updated the locale & languages that trigger the auto-allowing ads on YT
- Added support for showing 'local' On-Page Dialogs after the triggering of the auto-allowing ads on YT
- Updated the On-Page Dialogs code to allow them to be triggered by the extension (without an IPM ID, etc.)
- Updated some of the On-Page Dialogs modules to be more aligned with the same code in ABP (primarily in the `\onpage-dialog\background` and `\onpage-dialog\shared` folders)
- Updated some of the IPM modules to be more aligned with the same code in ABP (relating to the usage of `licenseStateList` and the use of `isDomainList` in the new `src/core/url` module).
- Added the `src/core/url` module
- Updated the TypeScript definition for the `ext` namespace, adding `Pages`

## Testing hints
The testing hints for the original [MR](https://gitlab.com/adblockinc/ext/adblock/adblock/-/merge_requests/300) still apply to this MR.

Since the code changes in this MR impact the IPM commands, please regression test the On Page Dialogs and New Tabs commands.

Additional testing hints:

The following preference can be updated to a locale code to allow or prevent the showing of the On-Page Dialog on YT after allowing the listing.  The locale code(s) should be in addition to the locale codes on the `yt_allowlist_language_codes`.  In other words, the locales on the `yt_allowlist_with_dialog_language_codes` preference will be considered as well as the ones in `yt_allowlist_language_codes`, so they don't need to be duplicated in that preference 

`Prefs.set('yt_allowlist_with_dialog_language_codes', ['en']);`

1. Since we're using the 'pause once' feature, verify that YT is no longer allow listed when:

* the user navigates away (closes any/all of the tabs on YT or shut down the browser)
* navigates back to YT

2. The current log messages ("yt-site.ad-wall-detected", "yt-site.navigation") continue to be sent for all users no matter the language
3. The hard end date works as designed - the auto-allow listing feature is no longer invoked after the hard end date.

See eyeo/extensions/extensions!10