feat: Add Button Component to UI Components, Part 1 [EXT-79]
### In This MR

This MR contains a preponderance of the code needed for adding components to the `ui-components` directory and contains enough of an example that other should be able to work from it. 

It:

- expands the folder structure 
- adds testing infrastructure and examples
- adds button component

To unblock other people, I think we could go ahead and merge this and do a follow-on MR containing the outstanding tasks.  

### Proposed for Followup

These tasks are:
- adding hover/active states
- refining themes: current values for some are incorrect and result in illegible text
- adding type annotations
- seeing how class concatenation might benefit from `clsx`, especially with the `colorOverrides`  — I am not against it but have not had a chance to look yet

## How Can I Experience This For Myself?

In `ui-components`, run `npm i && npm run dev` and look at the silly website that shows the buttons in addition to theme colors. The code for this also shows how the themed variables become integrated into the code via Tailwind.

## Can You Show Me a Picture?

![Screenshot_2024-08-28_at_19.28.29](/uploads/a3cdad8c06a41674c435366c762e6e4b/Screenshot_2024-08-28_at_19.28.29.png)

See eyeo/extensions/extensions!58