chore: Get rid of the explicit core dependency, get it through the sdk instead [noissue]
I noticed that AdBlock was depending on `adblockpluscore` from
npm. However, in Jan 2023 we merged the `adblockpluscore`
codebase into the `webext-ad-filtering-solution` codebase and
stopped publishing new versions to NPM.

The file that AdBlock was importing from `adblockpluscore` still
exists inside the SDK, and is being exported, so we can trivially
remove this dependency.

The only change from the old version of `contentTypes` being
imported to the new version being imported is that the new one
has support for the `webbundle` type. This is an irrelevant
change, since Google discontinued work on Web Bundles in Feb
2023, immediately after we added support for blocking them.

See eyeo/extensions/extensions!75