fix: Refactor the `getting` of the device ID [EXT-305]
The changes in this MR fix the problem described in the Jira ticket.  

The changes include:
- Generating the device ID the first time the `getUserId()` is invoked and caching the user ID in a local variable.
- Returning the local variable when it's set

This avoids the race condition at start-up when `getUserId()` is invoked multiple times.  It also improves the performance of the extension by removing a get from persistent storage every time the `getUserId()` is invoked.

## Testing hints:

In addition to the scenario described in the Jira ticket, the device ID in the following items should be verified. The device ID should remain the same (consistent) as the one in the URL of the first run page.

* After a browser or extension restart, the device ID should remain the same
* Existing installations should retain the same device ID (a new device ID shouldn't be created when this fix is applied)

See eyeo/extensions/extensions!94