test: Standardize default .env format for e2e tests [Refs EXT-222]
For extension secrets, we have a .env file which is not committed, and a
.env.defaults which has any default values and shows the format.

For the e2e tests in ABP, we had a .env.e2e file which is not committed, and a
.env.e2e.template which contains defaults and shows the format, but needs to be
copied for the defaults to take effect.

This change renames .env.e2e.template to .env.e2e.defaults, and has the process
that uses it load the defaults file as a fallback. This means that it no longer
needs to be copied if secrets aren't being added.

See eyeo/extensions/extensions!99