test: Fix local run on Adblock end-to-end tests from non-US locale [noissue]
I'm trying to investigate the flakiness of these end-to-end
tests.

While doing so, I have found two small issues which specifically
make it problematic for me to run the end-to-end tests on my
computer.

1. The install test assumes a locale of `en-US`. I've updated the test to allow
   any locale here.
2. Downloading the browser was done with a timeout, and I found that this meant
   sometimes the tests would fail to start if I couldn't download the browser
   that fast. This is especially a problem when running the tests in a docker
   container since the browser isn't cached then. I've fixed this by putting the
   browser download in its own `before` block with the timeout disabled.

This unfortunately does not yet address the actual flakiness I've been seeing on
the pipelines, but it's a start.

See eyeo/extensions/extensions!107