test: Increase the helper extension timeout [EXT-245]
It increases the helper extension timeout. The hypothesis is that when the helper extension tries to open the options page, under stress conditions the extension is not ready for that to happen.

This is only an attempt to fix the issue, given that's very hard to reproduce locally.
We'll keep an eye on the main pipeline runs to see if it shows up again.

See eyeo/extensions/extensions!108