test: Refresh page when elements are not yet rendered in options page [EXT-298]
This MR is fixing massive flakiness we observe on AdBlock pipelines.

Main reason was already evaluated by Ursa in the[ Jira ticket](https://eyeo.atlassian.net/browse/EXT-298). TL;DR: we don't wait for options to be rendered before starting to do actions, therefore we end up in a state where options are empty and we cannot interact with them:

![allowlists_websites.png](/uploads/dab36e1cf23042d8690e9254da31b2e8/allowlists_websites.png)

This happens only when automation is performing actions.

As suggested by Ursa, the solution is to refresh the page to allow the extension to render it again.

I've added console log to indicate when this mechanism is used and current execution will look as follow: 

![Screenshot 2024-09-24 at 12.53.53.png](/uploads/97ec764ff565dd93d40ab729fd8c4cd3/Screenshot_2024-09-24_at_12.53.53.png)

This is only a workaround as we will have an event to inform tests when we can interact with the options page: https://eyeo.atlassian.net/browse/EXT-335.

Note: for \`disables allow AA\` I had to introduce parameter to not "forceRefresh" the page when looking for element. In this scenario we are waiting for message that is shown after disabling AA. When we refresh the page we loose it, therefore tests fail.

See eyeo/extensions/extensions!114