feat: Auto allow list YT after the ad wall is detected [EXT-333]
TODO: Remove the locale messages for the Dialog?

The changes in this MR remove the locale and date checks used to test the various approaches to reducing user churn. These changes will now cause AdBlock to 'auto allow list' YT for all users when the YT ad wall is detected.

The following describes the functionality that will be in place with these changes:

1. When a user installs or receives the extension update, the extension captures that date as the start date. I kept this meta data because it might be helpful in the future. I can remove it if we feel it isn't needed.
2. When any user navigates to YouTube, AdBlock will continue to:
   * send a log event message "yt-site.navigation" 
3. When any user navigates to YouTube, **_and_** the YouTube ad wall appears, AdBlock will:
   * Detect the ad wall using the selector "ytd-enforcement-message-view-model"
   * Send a log event message "yt-site.ad-wall-detected" 
   * Allow list YT on that tab (pause once)
   * Send a log event message "yt-auto.allowlisted"
   * Reload the tab
   * Note: the "yt-site.navigation" event message will be sent again since the tab is reloaded
4. Remove the testing-specific Preferences:
   * 'yt_allowlist_language_codes'.
   * 'yt_allowlist_hard_end_date'. 


## Testing hints:

If YouTube isn't displaying its ad wall, the following snippet can be run on the console tab of the developer tools for the YT tab. This snippet will trigger the items in step 3 above. This code snippet must be run in the browser developer console, with the 'AdBlock' extension context selected in the upper left corner of the console tab.

`browser.runtime.sendMessage({ type: "yt-site.ad-wall-detected", userLoggedIn: false });`

Suggested test cases:

1. Since we're using the 'pause once' feature, verify that YT is no longer allow listed when:

* the user navigates away (closes any/all of the tabs on YT or shuts down the browser)
* then navigates back to YT

2. The log messages ("yt-site.navigation", "yt-site.ad-wall-detected", "yt-auto.allowlisted" ) are sent for all users

See eyeo/extensions/extensions!113