fix: Removed YouTube mention from extension description [EXT-378]
Reverted extension description to previous version but without mentioning YouTube.

Note that the fix is based on the previous description, since there wouldn't be much left if we remove YouTube from the current description (i.e. `Remove ads everywhere you browse.`, so the thinking is that the old description would be a bit more helpful for this purpose.
- Current description: `Remove ads on YouTube and everywhere else you browse.`
- Previous description: `Block YouTube™ ads, pop-ups & fight malware!`

Release notes were added under a new "Unreleased hotfix" section to clearly differentiate them from changes that wouldn't be part of the hotfix release.

- [x] I have updated the release notes

---

- [x] My changes didn't require a documentation update

---

- [x] My changes didn't require new tests

---

- [x] All new and existing tests passed

See eyeo/extensions/extensions!130