test: Standarise end to end tests scripts across extension hosts [Refs EXT-352]
## Description

This MR standardises scripts commands to run end-to-end tests.

It's still a draft MR, but in a state where you can check it out, play with it and give feedback on commands that got introduced.

## Motivation and Context

We need to have standard commands to allow us to run `npm run test:end-to-end` and despite different frameworks they should perform end to end tests on a specific extension. This is needed for Lerna to be able to work properly.

## Checklist:

- [x] Developer can run `npm run test:end-to-end {chrome|edge|firefox} {2|3}` in both extensions directories and tests are triggered accordingly
- [x] Developer can use `grep` and other mocha command line options when running tests locally on AB
- [x] Developer can use `FORCE_HEADFUL` flag when running tests locally on AB
- [x] My changes doesn't require update of the release notes
- [x] My changes require a documentation update and I have updated the documentation accordingly
- [x] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!132