test: call afterSequence with fallback url inside beforeSequence [EXT-252]
## Context

All example failures happened in the `before` hook of the first suite running, which makes me think it's a flaky issue related with the state of the browser right after the extension has been installed. 

## Description

The fix in this MR is based on that assumption. Changes:

- Calls `afterSequence` with fallback url inside `beforeSequence`. Please see the details in the added comment in the code changes
- Reorders the `filterlists` suite to make smoke tests run before the filterlists tests. (Unrelated, but nice to have)

## Motivation and Context

https://eyeo.atlassian.net/browse/EXT-252

## Checklist:

- [X] My changes didn't require a documentation update
- [X] My changes didn't require new tests
- [X] All new and existing tests passed

See eyeo/extensions/extensions!146