build: Optimize host webpack is called within the hosts' gulp scripts [noissue]
## Description

This change alters how we call webpack from calling webpack once per bundle, to calling it once with multiple configurations. This improves the build time dramatically.

## Motivation and Context

While experimenting with Lerna, I was getting frustrated that the extensions take a long time to build.

I discovered that the majority of the time is spent running webpack repeatedly. In previous experience with the SDK, we found that webpack has a heavy start-up time, and so it is preferable to start it up only once.

This change would reduce both CI time spent in the "build" tasks, and dev time building the extensions. For example, see [this AdBlock build pipeline from this change](https://gitlab.com/eyeo/extensions/extensions/-/jobs/7986836047) which took 3:07, versus [this pipeline from the main branch](https://gitlab.com/eyeo/extensions/extensions/-/jobs/7986489930) which took 5:36.

## Checklist:

- [x] My changes do not require release notes
- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!148