tests: Run chromium instead of Chrome in ABP local and docker end to end tests [EXT-352]
## Description

This MR is standardising the local run of e2e to be aligned in terms of the MV3 chromium browser that is tested in both Hosts. 

## Motivation and Context

As suggested [here](https://gitlab.com/eyeo/extensions/extensions/-/merge_requests/132#note_2139688768 "test: Standarise end to end tests scripts across extension hosts [Refs EXT-352]") we should standarise browsers we run end to end tests on.

Now we run on Chromium in AdBlock and on Chrome in AdBlock Plus, which might lead to confusion when running standardised commands for tests and break the flow for running tests in all workspaces at once.

This MR is changing LOCAL and Docker run of e2e tests to use Chromium instead of Chrome. 

Note: LambdaTests is still using Chrome because it doesn't support chromium as browser for tests: https://www.lambdatest.com/list-of-browsers. 

## Checklist:

- [x] My changes require a documentation update and I have updated the documentation accordingly
- [x] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!145