build: Feed webpack builds in a deterministic order [noissue]
## Description

The bundles passed to webpack are now always passed in the same order.

This required bringing in a different stream merging library, `merge2`, which can merge the streams in a deterministic order.

## Motivation and Context

In !148, we changed the build pipeline to have one webpack stream, instead of many. However, a side effect of this was that webpack might encounter the bundles in a different order. When this happens, it might assign different internal IDs to things that can result in the built artifact not being byte-for-byte identical between runs.

We have a test pipeline that makes sure that we can build an identical extension from the source bundles, so this pipeline was failing intermittently due to these non-deterministic IDs.

## Checklist:

- [x] My changes do not require release notes
- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests. I have run the existing tests multiple times, and spot checked the artifacts generated between multiple pipeline runs.
- [x] All new and existing tests passed

See eyeo/extensions/extensions!149