feat: Adds InterpolateTranslate component [no-issue]
## Description

This adds an `InterpolateTranslate` component, which is analogous to the [translateWithTags](https://gitlab.com/eyeo/extensions/extensions/-/blob/main/host/adblock/adblock-betafish/button/utils.js#L104) functionality that was built for the web components rewrite. It is a basic stub that should work in Adblock as it is and may require some changes as it is rolled out more widely.

It is to be used in cases where we have a translation that needs an HTML component integrated into it, like a link, a styled `span` (used in Adblock for block counts), or an `em` tag. For cases where this is not needed, the host's translate function can be used directly.

The component expects a translation function to be passed to it, so it can be used in both extensions. It is based on Adblock interpolations of translations. From a quick review, it seems like ABP has a less flexible interpolation built into their translations, so we will want to have the same kind of divider for translations once we add components there.

## Motivation and Context

As we move into adding new UI framework functionality when we add new UI, I anticipate we will need this component fairly early on, and since I still have the context from before, I have added it here. This is important because the imperative version we used did not work well with web components, as it runs once everything is rendered and is not as modular as we need / does not handle conditionally displayed content.


## Checklist:

<!-- Go over all the following points, keep those that apply and mark them with an `x` whenever the task is completed -->

- [ ] I ~~have updated the release notes~~ I do not believe this is necessary as it is an internal change

---

- [ ] My changes require a documentation update and I have updated the documentation accordingly
      <br>_\- or -_
- [x] My changes didn't require a documentation update

---

- [x] I have added tests to cover my changes
      <br>_\- or -_
- [ ] My changes didn't require new tests

---

- [x] All new and existing tests passed

See eyeo/extensions/extensions!144