build: Setup for ui-components and React in ABP [EXT-422]
<!-- Provide a general summary of your changes in the Title above -->
<!-- If this is a draft: add "Draft:" to the Title. -->

## Description

This MR adds support for React code inside ABP and also should enable use of `ui-components`. The configuration is commented at the moment to not affect the current code until we actually ship some react code in ABP.

Because of the dependency on `ui-components` I had to change the CI configuration a bit. It should be changed back when we will have setup for Lerna. 

**IMPORTANT:** If we will ship some React code before Lerna is ready, locally we will have to build first `ui-components` and after that the extensions.

## Motivation and Context

This is part of the on going effort of moving towards React of our UI.

<!-- Why is this change required? What problem does it solve? -->
<!-- If it fixes an open issue, please link to the issue here. -->
<!-- If this is a draft: Explain the reasoning for it to be a draft -->
<!-- and what input you are looking for to get from your teammates -->

## Checklist:

<!-- Go over all the following points, keep those that apply and mark them with an `x` whenever the task is completed -->

- [ ] I have updated the release notes

---

- [ ] My changes require a documentation update and I have updated the documentation accordingly
      <br>_\- or -_
- [x] My changes didn't require a documentation update

---

- [ ] I have added tests to cover my changes
      <br>_\- or -_
- [x] My changes didn't require new tests

---

- [x] All new and existing tests passed


Closes EXT-422

See eyeo/extensions/extensions!162