test: Moved MV3 and MV2 LambdaTest jobs into a single job [EXT-351]
Moves MV3 and MV2 LambdaTest jobs into a single job.

## Description

Configures the conf files and the yml in order to have only one LambdaTest job that runs both MV3 and MV3 test cases. Adds all necessary changes throughout the code.

## Motivation and Context

Currently, the MV3 and MV2 LambdaTest jobs are separate in the CI. This leads to overhead when trying to execute all E2E test cases: if starting both jobs at the same time, it can lead to failures; whereas, the alternative is to wait for a job to be complete and then start the other one. Additionally, this causes problems for the nightly runs, where they both start at the same time.

See eyeo/extensions/extensions!152