feat: add events for when an allowlist expired or extended [EXT-376]
In order to gather data from the Adblock extension when a smart allowlisting has expired or it expiration date was extended, we need to send data on the 2 new events created on the SDK under [this MR](https://gitlab.com/eyeo/adblockplus/abc/webext-ad-filtering-solution/-/merge_requests/1120 "feat: add events for when an allowlist expired or extended [EXT-376]"):

These new collected events will also send the default expiration window so that later on we can use this data to find the best window.

#### What to test:

The 2 new added events are sent to the `https://log.getadblock.com/v2/record_log.php` as follows:

* `allowlisting_expired` is sent when a smart allowlisting filter has expired. The event should have the following payload:

  ```javascript
  { 
    "event": "allowlisting_expired", 
    "payload": {
      "u": "b6m435ev65578701",
      "f": "E",
      "o": "Mac OS X",
      "l": "en-US",
      "t": "general",
      "v": "6.9.3",
      "allowlistDuration": 10000 
    } 
  }
  ```
* `allowlisting_renewed` is sent when a smart allowlisting filter has been extended/renewed. The event should have the following payload:

  ```javascript
  { 
    "event": "allowlisting_renewed", 
    "payload": { 
      "u": "b6m435ev65578701",
      "f": "E",
      "o": "Mac OS X",
      "l": "en-US",
      "t": "general",
      "v": "6.9.3",
      "allowlistDuration": 10000
     }
  }
  ```

#### How to test:

In order to test this, it is easier to create a custom filter that has an expiration time smaller than the default one. This will avoid having to wait a long time to check that the events were sent.

1. Set the extending for the filter

   ```javascript
   const autoExtendMS = 1000 * 60 // set the extending period to 1 minute
   ```
2. Create the filter's metadata

   ```javascript
   const metadata = {
       autoExtendMS,
       created: Date.now(),
       expiresAt: Date.now() + autoExtendMS, // set when the filter should expire
       origin: "popup",
   }; 
   ```
3. Create the smart allowlisting filter

   ```
   await ewe.filters.add(["@@example.com$document"], metadata); // change filter text here as appropriate
   ```

See eyeo/extensions/extensions!141