ci: Be more specific in "always" build rules [noissue]
## Description

This reviews the use of the "always" rule in our CI pipelines.

## Motivation and Context

Generally, when we've been saying a step should "always" run,
what we actually mean is that it should run if its previous steps
have succeeded. That is, we intend to use "on_success", instead
of "always", which actually means to run that step regardless of
how the previous steps went.

For example, in this
pipeline (https://gitlab.com/eyeo/extensions/extensions/-/pipelines/1494284091),
the build step failed. We would normally not expect the compliance tests to run
in this scenario, since they depend on the artifact from the build step.

## Checklist:

- [x] I have updated the release notes
- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!169