test: Continued test execution after installed page timeout [EXT-479]
Continued test execution after installed page timeout

## Description

When running our test cases on LambdaTest, the installed page is sometimes not displayed. For example, this test: LambdaTest Automation  

It might be a LambdaTest issue, might be caused by the helper extension, or something else.

This change gracefully handles that 50 second timeout and tries to run the test even if the installed page did not open. This leads to more stability and less failed test cases.

The new behavior will not apply to the smoke tests, because they specifically expect the installed page to be displayed, that's why an exclusion logic was created for this.

## In Practice

Here is a test execution result, before applying this change:

![Screenshot_2024-10-17_at_15.13.45](/uploads/422397c8ca71f6b43ca41b94432af0ca/Screenshot_2024-10-17_at_15.13.45.png)

And here is one, after applying the change:

![Screenshot_2024-10-18_at_14.05.07](/uploads/ea150b45152fdb0cfab87ff14de4fcd5/Screenshot_2024-10-18_at_14.05.07.png)

See eyeo/extensions/extensions!177