build: Automatic Gitlab release creation based on tag [EXT-461]
## Description

This adds a job that will automatically create a Gitlab release when the appropriate tag is committed. This is part of efforts to improve our current release process. 

This also adds a release-utils package that comes with a bunch of tests. The main point is to have a script that can get the correct release notes for a host and a version. 

## Checklist:

- [x] My changes require a documentation update and I have updated the documentation accordingly.
- [x] I have added tests to cover my changes
- [x] All new and existing tests passed

See eyeo/extensions/extensions!204