test: Make "displays acceptable ads" test working with locally hosted page [Refs EXT-497]
We had a (skipped) test that used Ecosia website AA. This
was unreliable and flaky.

## Description

Now we work with with locally hosted page that uses existing
Easylist and AA filters to detect whether AA is shown. The test
is skipped for Firefox for now (to be done in a separate MR).

## Motivation and Context

<!-- Why is this change required? What problem does it solve? -->
<!-- If it fixes an open issue, please link to the issue here. -->
<!-- If this is a draft: Explain the reasoning for it to be a draft -->
<!-- and what input you are looking for to get from your teammates -->

## Checklist:

- [x] My changes didn't require to update the release notes
- [x] I have updates the tests to cover my changes

---

- [ ] All new and existing tests passed

See eyeo/extensions/extensions!189