test: Run end to end tests for `Display Acceptable Ads` with local page on Firefox [EXT-497]
This MR: 
- adds ability to run localhost testpages as `testpages.adblockplus.org` in webDriverIO and selenium tests in Firefox 
- unskips `Display Acceptable Ads` for local run in Firefox in both `hosts`
- I had to change expected `acceptableAdsIsOn` to be always expected to be `true` in ABP. Based on the ticket description (and extension behaviour) it's always true (in FF and Chromium in ABP) 

Ability to run localhost testpages in Firefox is done through firefox profiles and proxy settings for those profiles. Proxy settings are read from `PAC` file that is hosted locally on test-server. 
Proxy Auto Config (PAC) scripts allow custom routing of requests. This PAC file is loaded when first page is opened by tests (and can be observed through Browser Console).

Using a PAC file in Firefox in WebdriverIO leads to unexpected behavior: opening each tabs in a new window (explanation why in the comment in MR). 

To improve that, I added a setting to prevent a new window from opening.

If the new window is needed for testing purposes, we can still force it through a new instance of the browser:

```
it('should open links in a new window for this test', async () => {
    const newBrowser = await browser.newSession({
        capabilities: {
            browserName: 'firefox',
            'moz:firefoxOptions': {
                prefs: {
                    'browser.link.open_newwindow': 2, // Open in a new window
                },
            },
        },
    });
    await newBrowser.url('https://example.com');
});
```

For selenium most of the work was done in get-browser-binary. This MR upgrades package to bring this features and uses `proxy` argument to start browser with proxy.

See eyeo/extensions/extensions!206