feat: Handle and report too many commands error [IM-165]
## Description

We implemented the capability to both extensions to handle multiple commands being sent by the IPM server during one ping response. However, we weren't sending back an error message in case the server sent more commands than allowed. This MR adds the capability for the extensions to send events to the IPM server that are not connected to a specific command, and makes use of that functionality to send the error message about having received too many commands.

~~I also needed to upgrade Babel because I got the following error when executing `npm run test`:~~

> ~~`[BABEL] […]/extensions/host/adblockplus/mocks/js/jest-polyfill.js: Requires Babel "^7.22.0 || >8.0.0-alpha <8.0.0-beta", but was loaded with "7.18.9".`~~

## Links

* JIRA issue: https://eyeo.atlassian.net/browse/IM-165
* Error event spec: https://eyeo.atlassian.net/wiki/spaces/B2C/pages/397410320/Draft+Too+Many+Commands+Error+Event
* Multiple commands spec: https://eyeo.atlassian.net/wiki/spaces/B2C/pages/245694615/Draft+Sending+multiple+commands
* IPM group ticket: https://eyeo.atlassian.net/browse/IM-127

## Checklist:

- [x] My changes require a documentation update and I have updated the documentation accordingly
- [x] I have added tests to cover my changes
- [x] All new and existing tests passed

See eyeo/extensions/extensions!163