fix: Fix checkboxes in component library and adblock [noissue]
## Description

This MR fixes the UI checkbox so that it highlights on keyboard navigation, which involved changes in the component library and Adblock. I checked ABP and the highlighting appears to work already.

It also fixes an issue where the color of the checkmark in monochrome themes was the same color as the background. 

## Checklist:

<!-- Go over all the following points, keep those that apply and mark them with an `x` whenever the task is completed -->

- [-] I have updated the release notes >> fixes internal bug
- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!226