[build] Upload the ABP mv3 source maps to Sentry [EXT-439]
In this MR we upload the Adblock Plus MV3 web extension source maps to Sentry
to be able to track the actual source code line for the reported errors.

## Description

This MR is introducing sourcemaps upload to sentry to our pipeline builds so we could debug better errors we get in sentry. 

To Do Before Merge: 
- [x] add README if we want to allow to upload sourcemaps locally from the machine
- [x] `.gitlab-ci.yml` config to trigger upload of sourcemaps on release
- [x] configure gitlab for sentry authentication (existing `SENTRY_AUTH_TOKEN`)

To Do After Merge:

- [ ] check the source maps are uploaded during the actual release

## Checklist:

- [x] I have updated the README file
- [x] My changes didn't require new tests

See eyeo/extensions/extensions!241