test: Fix flaky updating a single filter list test [EXT-581]
## Description

When updating filter lis(s) on MV3, the 'Last updated' status is stuck in 'Updating', instead of being 'Just now' after the update is finished, which causes flakiness in our tests. (See opened issue: https://eyeo.atlassian.net/browse/EXT-402). The correct status of 'Just now' is then displayed after the page is reloaded.

This MR adds a try/catch block to both of the tests that update the filterlists, so that in case the 'Just now' text is not shown, the page is reloaded before we check for the correct status again.

## Checklist:

- [x] My changes didn't require a documentation update
- [x] My changes didn't require new tests
- [x] All new and existing tests passed

See eyeo/extensions/extensions!258